pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-intelligence-core"
version = "0.2.0"
description = "Observable, file-backed document intelligence: ingest → index (FAISS) → deterministic graph."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "pymupdf>=1.24",
  "pymupdf4llm>=0.0.16",
  "pdfplumber>=0.11",
  "numpy>=1.24",
  "faiss-cpu>=1.7.4",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
pdf-ingest = "cli.ingest:main"
pdf-index = "cli.index:main"
pdf-graph = "cli.graph:main"
pdf-query = "cli.query:main"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["pdf_core*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.ruff]
line-length = 100
target-version = "py310"
